refactor/move presentation to presentation/web without requestDto
