--- app/presentation/views/Home.scala.html
@@ -1,13 +1,0 @@
-@*
- *
- * to do sample project
- *
- *@
-
-@(vv: presentation.views.model.ViewValueCommon)
-@common.Default(vv){
-   <ul>
-    <li><a>カテゴリー一覧</a></li>
-    <li><a>Todo一覧</a></li>
- </ul>
-}
