+++ app/presentation/web/views/Task.scala.html
@@ -1,8 +1,8 @@
 @import usecase.task._
 @import domain.model.task.Status
 @import helper._
-@import presentation.controllers._
-@(vv: presentation.views.model.ViewValueCommon, tasks: Seq[TaskItemDto], categories: Seq[ShowCategoryUseCaseDto], allState: Seq[Status])(implicit request: MessagesRequestHeader)
+@import presentation._
+@(vv: presentation.web.views.model.ViewValueCommon, tasks: Seq[TaskItemDto], categories: Seq[ShowCategoryUseCaseDto], allState: Seq[Status])(implicit request: MessagesRequestHeader)
 
 @nullCategoryName = @{"未分類"}
 @nullCategoryColor = @{"#000000"}
