+++ app/presentation/web/views/common/Default.scala.html
@@ -4,7 +4,7 @@
 * two arguments, a `String` for the title of the page and an `Html`
 * object to insert into the body of the page.
 *@
-@(vv: presentation.views.model.ViewValueCommon)(content: Html)
+@(vv: presentation.web.views.model.ViewValueCommon)(content: Html)
 
 <!DOCTYPE html>
 <html lang="ja">
