+++ app/presentation/web/views/task/Delete.scala.html
