+++ conf/routes
@@ -4,12 +4,12 @@
 # ~~~~
 
 # An example controller showing a sample home page
-GET     /                             presentation.controllers.HomeController.index()
+GET     /                             presentation.web.controllers.HomeController.index()
 
 # Map static resources from the /public folder to the /assets URL path
 GET     /assets/*file                 controllers.Assets.versioned(path="/public", file)
 
 # Task
-POST    /task                         presentation.controllers.TaskController.add()
-PUT     /task                         presentation.controllers.TaskController.update()
-DELETE  /task                         presentation.controllers.TaskController.delete()
+POST    /task                         presentation.web.controllers.TaskController.add()
+PUT     /task                         presentation.web.controllers.TaskController.update()
+DELETE  /task                         presentation.web.controllers.TaskController.delete()
